refactor: 변수명 수정, 디렉토리 생성
