+++ src/main/java/com/example/ReviewZIP/domain/sms/dto/request/SmsDto.java
@@ -1,4 +1,4 @@
-package com.example.ReviewZIP.domain.sms;
+package com.example.ReviewZIP.domain.sms.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
