+++ src/main/java/com/example/ReviewZIP/domain/token/RefreshTokenController.java
@@ -43,7 +43,7 @@ public class RefreshTokenController {
     @ApiResponses({
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200", description = "OK, 성공"),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "USER403", description = "비밀번호가 잘못되었습니다."),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "USER413", description = "이메일을 입력하지 않았습니다."),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "USER415", description = "이메일 또는 비밀번호를 입력하지 않았습니다."),
     })
     public ApiResponse<TokenDto> login(@RequestBody LoginRequestDto loginRequestDto) {
         return ApiResponse.onSuccess(refreshTokenService.login(loginRequestDto));
